requirements.txt
click==8.1.7
pandas==2.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
